test: resolve flaky repl test
